requirements.txt
py-cord
DateTime~=4.3
asyncio~=3.4.3
aiohttp==3.7.4.post0
python-dotenv==0.19.0
requests==2.26.0